feat : 공유 기능을 구현한다
